requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asyncer==0.0.7
asyncpg==0.30.0
attrs==25.1.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
chainlit==2.1.0
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
cloudinary==1.42.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.16.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Lazify==0.4.0
literalai==0.1.103
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
narwhals==1.24.1
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prisma==0.15.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.1
syncer==2.0.3
tenacity==9.0.0
timm==1.0.14
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.2
triton==3.2.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uptrace==1.29.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==0.20.0
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
